Split up tests more
