+++ products/jbrowse-web/src/tests/JBrowse.test.js
@@ -1,15 +1,8 @@
 import React from 'react'
 import '@testing-library/jest-dom/extend-expect'
 
-import {
-  cleanup,
-  fireEvent,
-  render,
-  waitFor,
-  getByRole,
-} from '@testing-library/react'
+import { fireEvent, render } from '@testing-library/react'
 import { toMatchImageSnapshot } from 'jest-image-snapshot'
-import { ErrorBoundary } from 'react-error-boundary'
 import { LocalFile } from 'generic-filehandle'
 import { TextEncoder } from 'web-encoding'
 import { clearCache } from '@jbrowse/core/util/io/RemoteFileWithRangeCache'
@@ -21,7 +14,6 @@ import JBrowseRootModelFactory from '../rootModel'
 import corePlugins from '../corePlugins'
 import * as sessionSharing from '../sessionSharing'
 import volvoxConfigSnapshot from '../../test_data/volvox/config.json'
-import chromeSizesConfig from '../../test_data/config_chrom_sizes_test.json'
 import { JBrowse, setup, getPluginManager, generateReadBuffer } from './util'
 import TestPlugin from './TestPlugin'
 jest.mock('../makeWorkerInstance', () => () => {})
@@ -35,8 +27,6 @@ global.Blob = (content, options) => ({ content, options })
 expect.extend({ toMatchImageSnapshot })
 
 setup()
-
-afterEach(cleanup)
 
 const waitForOptions = { timeout: 15000 }
 
@@ -101,72 +91,6 @@ test('toplevel configuration', () => {
   expect(test2).toEqual('test works')
 })
 
-test('variant track test - opens feature detail view', async () => {
-  const pluginManager = getPluginManager()
-  const state = pluginManager.rootModel
-  const { findByTestId, findAllByTestId, findByText } = render(
-    <JBrowse pluginManager={pluginManager} />,
-  )
-  await findByText('Help')
-  state.session.views[0].setNewView(0.05, 5000)
-  fireEvent.click(await findByTestId('htsTrackEntry-volvox_filtered_vcf'))
-  state.session.views[0].tracks[0].displays[0].setFeatureIdUnderMouse(
-    'test-vcf-604452',
-  )
-  const feats1 = await findAllByTestId('test-vcf-604452', {}, waitForOptions)
-  fireEvent.click(feats1[0])
-
-  // this text is to confirm a feature detail drawer opened
-  expect(await findByTestId('variant-side-drawer')).toBeInTheDocument()
-  const feats2 = await findAllByTestId('test-vcf-604452', {}, waitForOptions)
-  fireEvent.contextMenu(feats2[0])
-  fireEvent.click(await findByText('Open feature details'))
-  expect(await findByTestId('variant-side-drawer')).toBeInTheDocument()
-}, 10000)
-
-test('widget drawer navigation', async () => {
-  const pluginManager = getPluginManager(undefined, true)
-  const state = pluginManager.rootModel
-  const { findByTestId, findByText } = render(
-    <JBrowse pluginManager={pluginManager} />,
-  )
-  await findByText('Help')
-  state.session.views[0].setNewView(0.05, 5000)
-  // opens a config editor widget
-  fireEvent.click(await findByTestId('htsTrackEntry-volvox_filtered_vcf'))
-  fireEvent.click(await findByTestId('htsTrackEntryMenu-volvox_filtered_vcf'))
-  fireEvent.click(await findByText('Settings'))
-  await findByTestId('configEditor')
-  // shows up when there  active widgets
-  fireEvent.mouseDown(
-    getByRole(await findByTestId('widget-drawer-selects'), 'button'),
-  )
-  fireEvent.click(
-    await findByTestId(
-      'widget-drawer-selects-item-HierarchicalTrackSelectorWidget',
-    ),
-  )
-  await findByTestId('hierarchical_track_selector')
-
-  // test minimize and maximize widget drawer
-  expect(state.session.minimized).toBeFalsy()
-
-  await findByTestId('drawer-minimize')
-  fireEvent.click(await findByTestId('drawer-minimize'))
-  expect(state.session.minimized).toBeTruthy()
-
-  fireEvent.click(await findByTestId('drawer-maximize'))
-  expect(state.session.minimized).toBeFalsy()
-
-  // test deleting widget from select dropdown using trash icon
-  expect(state.session.activeWidgets.size).toEqual(2)
-  fireEvent.mouseDown(
-    getByRole(await findByTestId('widget-drawer-selects'), 'button'),
-  )
-  fireEvent.click(await findByTestId('ConfigurationEditorWidget-drawer-delete'))
-  expect(state.session.activeWidgets.size).toEqual(1)
-}, 20000)
-
 describe('assembly aliases', () => {
   it('allow a track with an alias assemblyName to display', async () => {
     const pluginManager = getPluginManager()
@@ -202,37 +126,6 @@ describe('nclist track test with long name', () => {
   }, 20000)
 })
 
-describe('test configuration editor', () => {
-  it('change color on track', async () => {
-    const pluginManager = getPluginManager(undefined, true)
-    const state = pluginManager.rootModel
-    const { getByTestId, findByTestId, findByText, findByDisplayValue } =
-      render(<JBrowse pluginManager={pluginManager} />)
-    await findByText('Help')
-    state.session.views[0].setNewView(0.05, 5000)
-    fireEvent.click(await findByTestId('htsTrackEntry-volvox_filtered_vcf'))
-    fireEvent.click(await findByTestId('htsTrackEntryMenu-volvox_filtered_vcf'))
-    fireEvent.click(await findByText('Settings'))
-    await findByTestId('configEditor')
-    fireEvent.change(await findByDisplayValue('goldenrod'), {
-      target: { value: 'green' },
-    })
-    await waitFor(
-      () =>
-        expect(getByTestId('box-test-vcf-604452')).toHaveAttribute(
-          'fill',
-          'green',
-        ),
-      waitForOptions,
-    )
-  }, 20000)
-})
-
-// eslint-disable-next-line react/prop-types
-function FallbackComponent({ error }) {
-  return <div>there was an error: {String(error)}</div>
-}
-
 describe('test sharing', () => {
   sessionSharing.shareSessionToDynamo = jest.fn().mockReturnValue({
     encryptedSession: 'A',
@@ -255,33 +148,6 @@ describe('test sharing', () => {
     expect(url.value).toBe('http://localhost/?session=share-abc&password=123')
   })
 })
-
-test('404 sequence file', async () => {
-  console.error = jest.fn()
-  const pluginManager = getPluginManager(chromeSizesConfig)
-  const { findAllByText } = render(
-    <ErrorBoundary FallbackComponent={FallbackComponent}>
-      <JBrowse pluginManager={pluginManager} />
-    </ErrorBoundary>,
-  )
-  await findAllByText('HTTP 404 fetching grape.chrom.sizes.nonexist', {
-    exact: false,
-  })
-})
-
-test('wrong assembly', async () => {
-  console.error = jest.fn()
-  const pluginManager = getPluginManager()
-  const state = pluginManager.rootModel
-  const { findAllByText } = render(<JBrowse pluginManager={pluginManager} />)
-  const view = state.session.views[0]
-  view.showTrack('volvox_wrong_assembly')
-  await findAllByText(
-    'Error: region assembly (volvox) does not match track assemblies (wombat)',
-    {},
-    waitForOptions,
-  )
-}, 15000)
 
 test('looks at about this track dialog', async () => {
   const pluginManager = getPluginManager()
